requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5
